requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-dotenv==1.4.2
django-filter==24.2
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
idna==3.7
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg==3.2.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
